refactor: merge standalone and metasrv table metadata allocators
